+++ tests/test_serving.py
@@ -9,12 +9,12 @@
     :license: BSD, see LICENSE for more details.
 """
 import os
+import socket
 import ssl
+import subprocess
 import sys
 import textwrap
 import time
-import subprocess
-
 
 try:
     import OpenSSL
@@ -450,7 +450,6 @@ def test_multiple_headers_concatenated_per_rfc_3875_section_4_1_18(dev_server):
         from httplib import HTTPConnection
     else:
         from http.client import HTTPConnection
-
     conn = HTTPConnection('127.0.0.1', server.port)
     conn.connect()
     conn.putrequest('GET', '/')
@@ -470,3 +469,23 @@ def test_multiple_headers_concatenated_per_rfc_3875_section_4_1_18(dev_server):
     assert res.read() == b'a ,b,c ,d'
 
     conn.close()
+
+
+def can_test_unix_socket():
+    if not hasattr(socket, 'AF_UNIX'):
+        return False
+    try:
+        import requests_unixsocket  # noqa: F401
+    except ImportError:
+        return False
+    return True
+
+
+@pytest.mark.skipif(not can_test_unix_socket(), reason='Only works on UNIX')
+def test_unix_socket(tmpdir, dev_server):
+    socket_f = str(tmpdir.join('socket'))
+    dev_server('''
+    app = None
+    kwargs['hostname'] = {socket!r}
+    '''.format(socket='unix://' + socket_f))
+    assert os.path.exists(socket_f)